Make the ScopePathHelper a decorator
The scope has to be passed into the method anyway so it might as well be
a decorator.
